+++ src/config/constants.h
@@ -16,7 +16,7 @@ class Client;
 extern const char *           kDefaultPath;
 extern const std::string      kDefaultRoot;
 extern const int              kBacklogSize;
-extern const int              kBufferSize;
+extern const unsigned long    kBufferSize;
 extern const time_t           kTimeoutSec;
 extern const long             kNsecLoopDelay;
 extern const std::string      kEOL;
